movieApp/requirements.txt
Flask==1.0.2
google-cloud-firestore==0.31.0
